+++ rasterizeddb_core/src/core/table_ext.rs
@@ -244,7 +244,7 @@ pub(crate) async fn process_chunk_async(
 
                     if db_type != DbType::STRING {
                         let size = db_type.get_size();
-                        let memory_chunk = MEMORY_POOL.acquire(size);
+                        let memory_chunk = MEMORY_POOL.acquire(size as usize);
 
                         let mut data_buffer = unsafe { memory_chunk.into_wrapper() };
 
@@ -283,7 +283,7 @@ pub(crate) async fn process_chunk_async(
 
                         let chunk_slice = cursor_vector.vector.as_slice();
 
-                        let str_memory_chunk = MEMORY_POOL.acquire(str_length);
+                        let str_memory_chunk = MEMORY_POOL.acquire(str_length as usize);
 
                         let mut preset_buffer =
                             unsafe { str_memory_chunk.into_wrapper() };
